tests splits by cards. turnng potion test added. move counter pro fixed in game.js turningHand func
